code/requirements.txt
absl_py
flax
jax
matplotlib
ml_collections
numpy
objax
optax
torch
torchvision
tqdm